requirements.txt
about-time==4.2.1
absl-py==1.4.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
alive-progress==3.1.4
altair==5.1.1
altgraph==0.17.4
annotated-types==0.5.0
ansi2html==1.8.0
anyio==3.6.2
aplus==0.11.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
array-record==0.4.1
arrow==1.2.3
asgiref==3.7.2
astropy==5.3.2
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
bayespy==0.5.26
beautifulsoup4==4.11.1
blake3==0.3.3
bleach==5.0.1
blinker==1.6.2
blis==0.7.10
boto==2.49.0
boto3==1.28.29
botocore==1.31.29
bqplot==0.12.40
branca==0.6.0
Brotli==1.1.0
cachetools==5.3.1
catalogue==2.0.9
certifi==2022.12.7
cffi==1.15.1
cftime==1.6.2
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.4
comtypes==1.1.14
confection==0.1.1
contourpy==1.1.0
cryptography==41.0.4
cycler==0.11.0
cymem==2.0.7
Cython==3.0.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2023.4.0
datasets==2.15.0
debugpy==1.6.5
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.8
Django==4.2.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dm-tree==0.1.8
dnspython==2.6.1
dropbox==11.36.2
editdistpy==0.1.3
emoji==2.8.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
entrypoints==0.4
et-xmlfile==1.1.0
etils==1.5.2
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.101.1
fastjsonschema==2.16.2
ffmpeg==1.4
ffmpeg-python==0.2.0
ffmpy==0.3.1
filelock==3.12.2
Flask==3.0.0
Flask-Ext==0.1
flatbuffers==23.5.26
folium==0.14.0
fontawesomefree==6.4.2
fonttools==4.40.0
fqdn==1.5.1
frozendict==2.3.8
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.3
gast==0.5.4
gensim==4.3.2
gitdb==4.0.10
GitPython==3.1.34
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.60.0
gradio==3.44.4
gradio_client==0.5.1
grapheme==0.6.0
graphviz==0.20.1
grpcio==1.59.0
gTTS==2.3.1
gzip-reader==0.1
h11==0.14.0
h2==4.1.0
h5py==3.9.0
hpack==4.0.0
html5lib==1.1
httpcore==0.18.0
httplib2==0.22.0
httptools==0.6.0
httpx==0.25.0
huggingface-hub==0.21.3
hyperframe==6.0.1
icecream==2.1.3
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
imbalanced-learn==0.11.0
imblearn==0.0
importlib-metadata==6.5.0
importlib-resources==6.1.0
iniconfig==2.0.0
ipydatawidgets==4.3.5
ipykernel==6.20.2
ipyleaflet==0.17.3
ipympl==0.9.3
ipython==8.8.0
ipython-genutils==0.2.0
ipyvolume==0.6.3
ipyvue==1.9.2
ipyvuetify==1.8.10
ipywebrtc==0.6.0
ipywidgets==8.1.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
json5==0.9.11
JsonForm==0.0.2
jsonpointer==2.3
jsonschema==4.17.3
JsonSir==0.0.2
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab==3.5.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.19.0
jupyterthemes==0.20.0
kaitaistruct==0.10
keras==2.14.0
kiwisolver==1.4.4
langcodes==3.3.0
lesscpy==0.15.1
libclang==16.0.6
libretranslatepy==2.1.1
lightning-utilities==0.10.1
llvmlite==0.40.1
locket==1.0.0
lxml==4.9.3
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.1
marshmallow==3.21.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
micropip==0.3.0
missingno==0.5.2
mistune==2.0.4
ml-dtypes==0.2.0
mongoengine==0.28.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
multipledispatch==1.0.0
multiprocess==0.70.15
murmurhash==1.0.9
mutagen==1.47.0
mysql==0.0.3
mysql-connector-python==8.0.31
mysqlclient==2.2.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.8
nbformat==5.7.3
nest-asyncio==1.5.6
netCDF4==1.6.3
networkx==3.1
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.57.1
numpy==1.26.0
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.9.7
outcome==1.2.0
packaging==23.0
pafy==0.5.5
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.0
pathvalidate==3.2.0
pathy==0.10.2
patsy==0.5.3
pefile==2023.2.7
pgmpy==0.1.23
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==4.1.0
playsound==1.3.0
plotly==5.16.1
pluggy==1.0.0
ply==3.11
portalocker==2.7.0
preshed==3.0.8
proglog==0.1.10
progress==1.6
progressbar2==4.2.0
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==12.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pybase64==1.3.0
pybbn==3.2.3
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pydeck==0.8.0
pydivert==2.1.0
pydot==1.4.2
pydotplus==2.0.2
pydub==0.25.1
pyerfa==2.0.0.3
pyflakes==3.2.0
pygame==2.5.2
Pygments==2.14.0
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
PyJWT==2.1.0
pymongo==4.6.2
Pympler==1.0.1
PyMuPDF==1.23.4
PyMuPDFb==1.23.3
PyMySQL==1.1.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
pypdf==3.5.0
PyPDF2==3.0.1
PyPDF3==1.0.6
pypiwin32==223
pyrsistent==0.19.3
pysat==3.0.6
PySocks==1.7.1
pyspark==3.5.0
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
Python-EasyConfig==0.1.7
python-json-logger==2.0.4
python-multipart==0.0.6
python-utils==3.7.0
python-vlc==3.0.20123
pythreejs==2.4.2
pyttsx3==2.90
pytube==15.0.0
pytube3==9.6.4
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pywaffle==1.1.0
pywin32==305
pywin32-ctypes==0.2.2
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
regex==2023.8.8
requests==2.28.2
requests-oauthlib==1.3.1
Resource==0.2.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rsa==4.9
s3transfer==0.6.2
scapy==2.5.0
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
selenium==4.14.0
selenium-wire==5.1.0
semantic-version==2.10.0
semver==3.0.1
Send2Trash==1.8.0
simpleaudio==1.0.4
six==1.16.0
sklearn==0.0.post5
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SpeechRecognition==3.9.0
speedtest-cli==2.1.3
sqlparse==0.4.4
srsly==2.4.7
stack-data==0.6.2
starlette==0.27.0
statsmodels==0.14.0
stone==3.3.1
streamlit==1.34.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.14.0
termcolor==2.3.0
terminado==0.17.1
textblob==0.17.1
tfds-nightly==4.9.3.dev202310060044
thinc==8.1.12
threadpoolctl==3.1.0
tinycss2==1.2.1
tkPDFViewer==0.1
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
toposort==1.10
torch==2.2.1+cu118
torch-tb-profiler==0.4.3
torchaudio==2.2.1+cu118
torchdata==0.7.1
torchmetrics==1.3.1
torchtext==0.17.1
torchvision==0.17.1+cu118
tornado==6.3.2
tqdm==4.64.1
traitlets==5.8.1
traittypes==0.2.1
translate==3.6.1
trio==0.22.2
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.10.0
tzdata==2023.3
tzlocal==4.3.1
ucimlrepo==0.0.3
uri-template==1.2.0
uritemplate==4.1.1
uritools==4.0.2
urlextract==1.9.0
urllib3==1.26.14
uvicorn==0.23.2
vaex==4.17.0
vaex-astro==0.9.3
vaex-core==4.17.1
vaex-hdf5==0.14.1
vaex-jupyter==0.8.2
vaex-ml==0.18.3
vaex-server==0.9.0
vaex-viz==0.5.4
validators==0.22.0
vboxapi==1.0
virtualenv==20.25.0
wasabi==1.1.2
watchdog==3.0.0
watchfiles==0.19.0
Wave==0.0.2
wcwidth==0.2.5
webcolors==1.12
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.4.2
websockets==11.0.3
Werkzeug==3.0.1
widgetsnbextension==4.0.8
wordcloud==1.9.2
wrapt==1.14.1
wsproto==1.2.0
xarray==2023.4.1
xxhash==3.4.1
xyzservices==2023.7.0
yarl==1.9.2
youtube-dl==2021.12.17
zipp==3.15.0
zlib-state==0.1.5
zstandard==0.21.0
